rm useless logic
